Remove unnecessary @testable
