+++ Tests/VaultCourierTests/Pkl+VaultCourierPayloads.swift
@@ -25,7 +25,7 @@ import Foundation
 #endif
 import PklSwift
 
-@testable import VaultCourier
+import VaultCourier
 
 extension IntegrationTests.Pkl {
     @Suite(.disabled())
@@ -59,21 +59,6 @@ extension IntegrationTests.Pkl {
 
             await #expect(throws: Never.self) {
                 _ = try await VaultToken.loadFrom(source: .url(url))
-            }
-        }
-
-        @Suite
-        struct VaultDuration {
-            @Test
-            func vault_duration_style_for_seconds() async throws {
-                let duration = Swift.Duration(secondsComponent: 5400, attosecondsComponent: 0)
-                #expect(duration.formatted(.vaultSeconds) == "5400s")
-            }
-
-            @Test
-            func vault_duration_style_for_hours() async throws {
-                let duration = Swift.Duration(secondsComponent: 5400, attosecondsComponent: 0)
-                #expect(duration.formatted(.vaultHours) == "1.5h")
             }
         }
     }
